fixed numeric type

+++ db/migrate/20140826173756_fix_approved_proposal_id_in_parking_spaces.rb
@@ -1,5 +1,5 @@
 class FixApprovedProposalIdInParkingSpaces < ActiveRecord::Migration
   def change
-    change_column :parking_spaces, :approved_proposal_id, :long
+    change_column :parking_spaces, :approved_proposal_id, :numeric
   end
 end
